requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
Werkzeug==2.3.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
